% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_crossterms.R
\name{all_crossterms}
\alias{all_crossterms}
\title{Create crossterms from two matrices}
\usage{
all_crossterms(Z_i, Z_j, make_sparse = TRUE)
}
\arguments{
\item{Z_i}{Matrix}

\item{Z_j}{Matrix}

\item{make_sparse}{Boolean for whether to output a sparse matrix.
Default is `TRUE`.}
}
\value{
Matrix of cross-terms between `Z_i` and `Z_j`.
}
\description{
A helper function for `G_generate` that produces cross-terms.
}
