% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_analytic.R
\name{bdiag_m}
\alias{bdiag_m}
\title{Fast block diagonal generator, taken from Matrix package examples}
\usage{
bdiag_m(lmat)
}
\arguments{
\item{lmat}{Matrix}
}
\value{
Block diagonal version of input
}
\description{
Copyright (C) 2016 Martin Maechler, ETH Zurich. Copied here for more
convenient integration.
}
