% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_analytic.R
\name{cov_organize_start}
\alias{cov_organize_start}
\title{Organize covariance matrices}
\usage{
cov_organize_start(cov_vec)
}
\arguments{
\item{cov_vec}{Character vector}
}
\value{
List of relevant outputs
}
\description{
Read the data frame of variance-covariance interactiosn and
organize the covariance matrices correctly.
Finds indices to feed into `cov_organize` function above
}
