% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_analytic.R
\name{eigenval_trim}
\alias{eigenval_trim}
\title{Trim eigenvalues and return a positive semidefinite matrix}
\usage{
eigenval_trim(V)
}
\arguments{
\item{V}{Matrix of dimension n x n}
}
\value{
A Matrix with negative eigenvalues removed
}
\description{
Removes the negative eigenvalues of the given matrix V, returning a positive
semidefinite (PSD) matrix.
}
