% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_analytic.R
\name{fbps_cov}
\alias{fbps_cov}
\title{Modified refund::fbps for covariance matrices}
\usage{
fbps_cov(
  data,
  subj = NULL,
  covariates = NULL,
  knots = 35,
  knots.option = "equally-spaced",
  periodicity = c(FALSE, FALSE),
  p = 3,
  m = 2,
  lambda = NULL,
  selection = "GCV",
  search.grid = T,
  search.length = 100,
  method = "L-BFGS-B",
  lower = -20,
  upper = 20,
  control = NULL
)
}
\arguments{
\item{data}{Data frame of values to fit}

\item{subj}{Character vector of subject IDs, defaults to `NULL`}

\item{covariates}{List of data points for each dimension; defaults to `NULL`}

\item{knots}{Numeric or list of number of knots or the vector of
knots for each dimension; defaults to 35.}

\item{knots.option}{Character knot selection method; defaults to
`"equally-spaced"`}

\item{p}{Numeric degrees of B-splines}

\item{m}{Numeric order of difference penalty}

\item{lambda}{User-selected smoothing parameters; defaults to NULL}

\item{selection}{Character selection of smoothing parameter; defaults to
"GCV:}

\item{search.grid}{boolean; defaults to `TRUE`. If false, uses `optim`.}

\item{search.length}{Integer no. of equidistant (log) smoothing parameters;
defaults to 100.}

\item{method}{See `stats::optim`; defaults to `"L-BFGS-B"`}

\item{lower, upper}{Numeric bounds for log smoothing parameter; defaults to
`-20, 20`. Passed to `stats::optim`.}

\item{control}{See `optim`.}
}
\value{
A smoothed matrix
}
\description{
Altered to speed up `refund::fbps`. Takes advantage of the symmetry of the
covariance matrix, i.e., lambda1 == lambda2
}
