% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_estimate.R
\name{G_estimate.fastFMM}
\alias{G_estimate.fastFMM}
\title{Estimate covariance of random components G(s1, s2), non-concurrent}
\usage{
\method{G_estimate}{fastFMM}(fmm, mum, betaHat, HHat, non_neg, MoM, silent)
}
\arguments{
\item{fmm}{"fastFMM" object.}

\item{mum}{Result from massive univariate step}

\item{betaHat}{Estimated functional fixed effects}

\item{HHat}{Estimate H(s)}

\item{non_neg}{Integer control for non-negativity calculation}

\item{MoM}{Controls method of moments estimator}

\item{silent}{Whether to print the step description during calculations.
Defaults to `TRUE`.}
}
\value{
An estimation of the G matrix
}
\description{
Estimates the covariance matrix G for random intercepts that occurs at Step 3
of the FUI method. Applies when `G_generate` cannot provide an analytic
solution.
}
\details{
A helper function for `var_analytic`.
}
