% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_estimate_randint.R
\name{G_estimate_randint}
\alias{G_estimate_randint}
\title{Special case of estimating covariance of random components G(s1, s2)}
\usage{
G_estimate_randint(data, L, out_index, designmat, betaHat, silent = TRUE)
}
\arguments{
\item{data}{A data frame containing all variables in formula}

\item{L}{Number of columns of outcome variables}

\item{out_index}{Indices that contain the outcome variables}

\item{designmat}{Design matrix of the linear models}

\item{betaHat}{Estimated functional fixed effects}

\item{silent}{Whether to print the step description during calculations.
Defaults to `TRUE`.}
}
\value{
An estimation of the G matrix
}
\description{
Estimates the covariance matrix G for random intercepts that occurs at Step 3
of the FUI method. A helper function for `fui`.
}
