% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_generate.R
\name{G_generate.fastFMM}
\alias{G_generate.fastFMM}
\title{Generate the G matrix in non-concurrent cases}
\usage{
\method{G_generate}{fastFMM}(fmm, mum, MoM)
}
\arguments{
\item{fmm}{"fastFMM" object. Ignored except for dispatch.}

\item{mum}{The massively univariate model output. Contains relevant data,
such as the transposed Ztlist from the univariate fit.}

\item{MoM}{indicator for type of MoM estimation}
}
\value{
A list with the new Z, the original Ztlist, and indices
}
\description{
Creates a design matrix from the massive univariate step, assuming the model
is not concurrent.
}
