% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_generate.R
\name{G_generate}
\alias{G_generate}
\title{Generic "G_generate" G matrix setup}
\usage{
G_generate(fmm, ...)
}
\arguments{
\item{fmm}{"fastFMM" object}

\item{...}{Additional arguments}
}
\value{
A design matrix for G estimation.
}
\description{
Creates the design matrix that allows for the estimation of the G matrix.
Output related to "G_estimate" dispatch.
}
