% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massmm.R
\name{massmm.fastFMM}
\alias{massmm.fastFMM}
\title{Massively univariate non-concurrent fit}
\usage{
\method{massmm}{fastFMM}(fmm, parallel, n_cores)
}
\arguments{
\item{fmm}{Object of "fastFMM" class or its inheritors.}

\item{parallel}{Boolean, passed from `fui`.}

\item{n_cores}{Number of cores for parallelization. Defaults to 1.}
}
\value{
A list containing results from the massive univariate step.
}
\description{
Fits separate models over the functional domain.
}
