% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massmm.R
\name{massmm.fastFMMconc}
\alias{massmm.fastFMMconc}
\title{Massively univariate concurrent fit}
\usage{
\method{massmm}{fastFMMconc}(fmm, parallel, n_cores)
}
\arguments{
\item{fmm}{Object of "fastFMM" class or its inheritors.}

\item{parallel}{Boolean, passed from `fui`.}

\item{n_cores}{Number of cores for parallelization. Defaults to 1.}
}
\value{
A list of outputs from `unimm`.
}
\description{
Fits separate models over the functional domain for concurrent models.
}
