% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massmm.R
\name{massmm}
\alias{massmm}
\title{Generic "massmm" model fit}
\usage{
massmm(fmm, ...)
}
\arguments{
\item{fmm}{An object that is or inherits from the "fastFMM" class.}

\item{...}{Additional arguments.}
}
\value{
Results depends on the dispatched method.
}
\description{
Fits the model over the functional domain based on the object class.
}
