% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massmm.R
\name{massmm_apply}
\alias{massmm_apply}
\title{Massively univariate parallelization helper}
\usage{
massmm_apply(fmm, parallel, n_cores = 1)
}
\arguments{
\item{fmm}{Object of "fastFMM" class or its inheritors.}

\item{parallel}{Boolean, taken from `fui` arguments.}

\item{n_cores}{Number of cores for parallelization. Defaults to 1.}
}
\value{
A list of fitted univariate models.
}
\description{
A helper function that handles the parallelization of the model fitting.
Helps to clean up the code of `massmm` executions. Dispatches `unimm`
based on the class of the provided univariate model object.
}
