% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massmm.R
\name{massmm_outs}
\alias{massmm_outs}
\title{Massively univariate list outputs}
\usage{
massmm_outs(mass_list, fmm)
}
\arguments{
\item{mass_list}{The output from `massmm_apply`}

\item{fmm}{Object of class "fastFMM" with parameters for model fitting.}
}
\value{
A list of betaTilde, the AIC matrix, and residuals (if applicable).
}
\description{
Helper function that returns basic outputs from `massmm_apply`.
}
