% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastFMM.R
\name{print.fastFMMconc}
\alias{print.fastFMMconc}
\title{Printing the fastFMMconc object}
\usage{
\method{print}{fastFMM}(fmm)
}
\arguments{
\item{fmm}{Object to print}
}
\value{
Basic list of entries
}
\description{
Simple method to look at the relevant fields.
}
