% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_analytic.R
\name{pspline_setting}
\alias{pspline_setting}
\title{Copy of pspline.setting.R from refund}
\usage{
pspline_setting(
  x,
  knots = select_knots(x, 35),
  p = 3,
  m = 2,
  periodicity = FALSE,
  weight = NULL
)
}
\arguments{
\item{x}{Marginal data points}

\item{knots}{The list of interior knots of the numbers of interior knots}

\item{p}{Degrees for B-splines, default = 3}

\item{m}{Orders of difference penalty, default = 2}
}
\description{
A slightly modified copy of [pspline.setting](https://rdrr.io/cran/refund/src/R/pspline.setting.R)
from `refund`. Copied here because the original function is not exported from
the package. Called within `fbps_cov` within `var_analytic`.
}
