% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_knots.R
\name{select_knots}
\alias{select_knots}
\title{select_knots.R from refund package}
\usage{
select_knots(t, knots = 10, p = 3, option = "equally-spaced")
}
\arguments{
\item{t}{Numeric}

\item{knots}{Numeric scalar or vector, the number/numbers of  knots or the
vector/vectors of knots for each dimension. Default = 10}

\item{p}{Numeric, the degrees of B-splines. Default = 3.}

\item{option}{Character, knot spacing, can be `"equally-spaced"` or
`"quantile"`}
}
\description{
Copied from [select_knots](https://rdrr.io/cran/refund/src/R/select_knots.R)
because the original is not exported for use.
}
