% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimm.R
\name{unimm.fastFMM}
\alias{unimm.fastFMM}
\title{Fit a univariate mixed model}
\usage{
\method{unimm}{fastFMM}(fmm, l)
}
\arguments{
\item{fmm}{A `fastFMM` object that contains parameters for the fit.}

\item{l}{location to fit the model}
}
\value{
a list containing point estimates, variance estimates, etc.
}
\description{
Fits a mixed model at location l. Part of Step 1 of the FUI approach.
}
