% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimm.R
\name{unimm.fastFMMconc}
\alias{unimm.fastFMMconc}
\title{Fit a univariate mixed model for concurrent models}
\usage{
\method{unimm}{fastFMMconc}(fmm, l)
}
\arguments{
\item{fmm}{A `fastFMM` object that contains parameters for the fit.}

\item{l}{location to fit the model}

\item{MoM}{indicator of type of MoM estimator. Coerced to `MoM == 1` to
prevent storage of all Z matrices.}
}
\value{
a list containing point estimates, variance estimates, etc. Also
contains `ztlist`, a list of transposed `Z` matrices.
}
\description{
Fits a mixed model at location l. Part of Step 1 of the FUI approach. Returns
the list of Z matrices for concurrent estimation.
}
