% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimm.R
\name{unimm_lmer}
\alias{unimm_lmer}
\title{Fit and return a univariate mixed model (helper)}
\usage{
unimm_lmer(formula, data, family)
}
\arguments{
\item{formula}{Model formula.}

\item{data}{Data frame of observations to fit}

\item{family}{Character, family of model}
}
\value{
an `lme4` model chosen between `lmer` or `glmer`, as appropriate.
}
\description{
Helper for `unimm`. Detects whether to use `lmer` or `glmer` and returns
the resulting model. Also helpful for producing a sample model during the
massively univariate step.
}
