% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimm.R
\name{unimm_outs}
\alias{unimm_outs}
\title{Get relevant features of univariate models}
\usage{
unimm_outs(fit_uni, fmm)
}
\arguments{
\item{fit_uni}{An `lme4` object corresponding to the fit at some point on the
functional domain.}

\item{fmm}{A "fastFMM" class object with parameters of the fit}
}
\value{
A list of relevant features of `fit_uni`
}
\description{
Helper for `unimm` that returns various qualities of the univariate fit
produced by `unimm_lmer` (also a helper).
}
