% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_analytic.R
\name{var_analytic}
\alias{var_analytic}
\title{Analytic variance calculation}
\usage{
var_analytic(
  fmm,
  mum,
  smoothed,
  MoM,
  non_neg,
  nknots_cov,
  seed,
  parallel,
  silent
)
}
\arguments{
\item{fmm}{Object of class "fastFMM".}

\item{mum}{Massively univariate model output of class "massmm"}

\item{smoothed}{Outputs from the smoothing Step 2}

\item{MoM}{integer, indicates type of MoM estimator (1 or 2)}

\item{non_neg}{Integer, indicates type of non-negativity calculation}

\item{nknots_cov}{Integer, number of knots for splines}

\item{seed}{Integer random seed}

\item{parallel}{Logical, whether to use parallel processing}

\item{silent}{Logical, suppresses messages when `TRUE`. Passed from `fui`.}
}
\value{
List of final outputs of `fui`
}
\description{
Helper for `fui`. Analytic calculation for Gaussian models.
}
