% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_parallel.R
\name{var_par_wrapper}
\alias{var_par_wrapper}
\title{Parallelization wrapper}
\usage{
var_par_wrapper(s, fmm, mum, Z, RHat, HHat, id_list, obs_ind, res_template)
}
\arguments{
\item{s}{Integer index for variance calculations}

\item{fmm}{A "fastFMM" object}

\item{mum}{Output from the massively univariate step}

\item{Z}{Matrix. This is only included to be consisntent with the
non-concurrent call and does nothing.}

\item{RHat}{Smoothed coefficients}

\item{HHat}{Smoothed variance coefficients}

\item{id_list}{List of unique IDs}

\item{obs_ind}{List of vectors of rows corresponding to each subject}

\item{res_template}{Index template}
}
\value{
List of entries at index `s` for XTV inverse, the variance of
beta-tilde, and trimmed H-hat.
}
\description{
To parallelize, the method needs to be dispatched properly. The normal
dispatch will pick the wrong argument, so we can use a wrapper. This
also has error handling capabilities to report the index of failure.
}
