% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_parallel.R
\name{var_parallel.fastFMM}
\alias{var_parallel.fastFMM}
\title{Parallel variance calculation for non-concurrent models}
\usage{
\method{var_parallel}{fastFMM}(fmm, mum, s, Z, RHat, HHat, id_list, obs_ind, res_template)
}
\arguments{
\item{fmm}{A "fastFMM" object}

\item{mum}{Output from the massively univariate step}

\item{s}{Integer index for variance calculations}

\item{Z}{Matrix}

\item{RHat}{Smoothed coefficients}

\item{HHat}{Smoothed variance coefficients}

\item{id_list}{List of unique IDs}

\item{obs_ind}{List of vectors of rows corresponding to each subject}

\item{res_template}{Index template}
}
\value{
List of entries at index `s` for XTV inverse, the variance of
beta-tilde, and trimmed H-hat.
}
\description{
Calculate the variance of coefficients with index s.
}
