% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_parallel.R
\name{var_parallel}
\alias{var_parallel}
\title{Parallel variance calculation}
\usage{
var_parallel(fmm, ...)
}
\arguments{
\item{fmm}{A "fastFMM" object}

\item{...}{Additional arguments}
}
\value{
Calculated variance at location s
}
\description{
A helper function for `var_analytic` that calculations the variance elements
of the covariance matrix.
}
