% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_estimate.R
\name{G_estimate}
\alias{G_estimate}
\title{Generic "G_estimate" dispatch}
\usage{
G_estimate(fmm, ...)
}
\arguments{
\item{fmm}{"An object that inherits from the "massmm" class"fastFMM object.}

\item{...}{Additional arguments}
}
\value{
An estimation of the G matrix
}
\description{
Estimates covariance of random components. The related function for the
random intercept flag (`G_estimate_randint`) does not require a generic
because the operations are the same for concurrent and non-concurrent models.
}
