% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_generate.R
\name{G_generate}
\alias{G_generate}
\title{Creates the design matrix that allows for estimation of G}
\usage{
G_generate(data, Z_lst, RE_table, MoM, ID = "id")
}
\arguments{
\item{data}{Data frame that contains the predictors and outcome}

\item{Z_lst}{Transposed list of Z matrices from the univariate fits}

\item{RE_table}{Table of random effects and interactions, generated from the
`lmerMod` object}

\item{MoM}{Integer to determine type of MoM estimation.}

\item{ID}{Name of the ID factor, assuming names of `HHat` are generated from
the same table in the same order}
}
\value{
List containing Z matrices and indices (unsure)
}
\description{
The function `G_estimate` uses a MoM method,
and `G_estimate_randint` is a special case of `G_estimate`.
}
\details{
Because `G_estimate`
}
