% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lick}
\alias{lick}
\title{Jeong et al. (2022) licking behavior data}
\format{
## `lick`
A data frame with 2120 rows and 94 columns:
\describe{
 \item{id}{Unique mouse ID}
 \item{session}{Session}
 \item{trial}{Trial}
 \item{lick_rate_050, lick_rate_100, lick_rate_150, lick_rate_200}{The average licking rate 0.5, 1.0, 1.5, and 2.0 seconds after sucrose administratuib}
 \item{iri}{The inter-reward interval relative to the previous lick}
 \item{photometry_1, ..., photometry_43}{Photometry recordings over the trial}
 \item{lick_1, ..., lick_43}{Whether a mouse was licking at a time point}
}
}
\source{
<https://dandiarchive.org/dandiset/000351/draft>
}
\usage{
lick
}
\description{
Data from Jeong H, Taylor A, Floeder JR, Lohmann M, Mihalas S, Wu B, Zhou M,
Burke DA, Namboodiri VMK. Mesolimbic dopamine release conveys causal
associations. Science. 2022 Dec 23;378(6626):eabq6740. doi:
10.1126/science.abq6740.
}
\details{
Data includes photometry recordings from an experiment where head-fixed mice
were ranodmly administered sucrose. Data is downsampled for convenience.
}
\keyword{datasets}
