% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimm.R
\name{unimm}
\alias{unimm}
\title{Generic "unimm" model fitting}
\usage{
unimm(fmm, ...)
}
\arguments{
\item{fmm}{An object that is or inherits from the "fastFMM" class.}

\item{...}{Additional arguments.}
}
\value{
Results depends on the dispatched method.
}
\description{
Dispatches class-specific methods for fitting a single point during the
univariate mixed model step.
}
\keyword{internal}
